+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Uninitialized max transaction size.
+- Missing validation of the number of command and address bits.
 
 ## [1.0.2] - 2024-08-29
 ### Fixed